backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
browser_use_sdk
python-dotenv==1.0.1